pyproject.toml
[tool.poetry]
name = "pygame-rgb-matrix"
version = "0.1.0"
description = "PyGame RGB Matrix Display Board"
authors = ["Louis King <jinglemansweep@gmail.com>"]
readme = "README.md"
packages = [{include = "src/app"}]

[tool.poetry.dependencies]
python = "^3.9"
poetry = "^1.3.2"
pygame = "^2.1.2"
black = "^23.1.0"
numpy = "^1.24.2"
python-dotenv = "^0.21.1"
pillow = "^9.4.0"
paho-mqtt = "^1.6.1"
aiohttp = "^3.8.3"
asyncio = "^3.4.3"
feedparser = "^6.0.10"
pygame-ft = {git = "https://github.com/jinglemansweep/pygame-ft.git"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
